requirements.txt
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gspread==6.2.1
oauthlib==3.3.1
